% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raerdata.R
\name{GSE99249}
\alias{GSE99249}
\title{RNA sequencing data from study GSE99249}
\usage{
GSE99249()
}
\value{
A list containing:
\itemize{
\item \code{bams} A \link{BamFileList} object, indicating the BAM file paths and
BAI indexes.
\item \code{fasta} A path to a fasta file from chr18 of hg38
\item \code{snps} a GRanges object containing known SNPs from the REDIportal
database (hg38)
}
}
\description{
Study
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE99249}{GSE99249}
examined RNA editing in WT, ADAR1KO, and ADAR1-p150 HEK293T cells treated
with and without interferon beta.
}
\details{
\code{GSE99249()} will download BAM and BAM index files from 6 RNA-seq
libraries. 3 libraries are ADAR1 knockout cells treated with interferon
beta and 3 libraries are wild type cells treated with interferon beta.
The BAM files contain alignments from chromosome 18.
}
\examples{
GSE99249()

}
