% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raerdata.R
\name{NA12878}
\alias{NA12878}
\title{Whole genome and RNA sequencing data from NA12878 cell line}
\usage{
NA12878()
}
\value{
A list containing:
\itemize{
\item \code{bams} A \link{BamFileList} object, indicating the bam file paths and BAI indexes.
\item \code{fasta} the path to a fasta file containing the first megabase of chr4
\item \code{snps} a GRanges object containing snps from the first megabase of chr4
}
}
\description{
Whole genome and RNA sequencing data from NA12878 cell line
}
\details{
Will download BAM and BAM index files from whole genome and RNA sequencing of
the NA12878 cell line, The data is from the first megabase of chromosome 4. Additionally
a fasta file and a database of known SNPs will be downloaded.
}
