% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raerdata.R
\docType{package}
\name{raerdata}
\alias{raerdata}
\title{raerdata}
\description{
A collection of datasets and databases to demonstrate RNA-editing
analysis approaches using the raer package.
}
\details{
\code{\link{atlases}} a collection of RNA editing databases

\code{\link{NA12877}} Whole genome and RNA sequencing from the NA12877 cell line

\code{\link{GSE99249}}  RNA sequencing data from a study that examined RNA editing in
WT, ADAR1KO, and ADAR1-p150 HEK293T cells treated with and without interferon beta.

\code{\link{pbmc_10x}}  single cell RNA sequencing data from human PBMCs from 10x Genomics
}
